Do some explicit conversions of number types.

+++ src/command_fileinfo.cpp
@@ -307,7 +307,7 @@ public:
         const auto buffers_capacity= static_cast<double>(info_handler.buffers_capacity / (1000 * 1000)) / 1000; // NOLINT(bugprone-integer-division)
 
         if (info_handler.buffers_capacity != 0) {
-            const auto fill_factor = std::round(100 * static_cast<double>(info_handler.buffers_size) / info_handler.buffers_capacity);
+            const auto fill_factor = std::round(100 * static_cast<double>(info_handler.buffers_size) / static_cast<double>(info_handler.buffers_capacity));
             std::cout << "  Sum of buffer capacities: " << info_handler.buffers_capacity << " (" << buffers_capacity << " GB, " << fill_factor << "% full)\n";
         } else {
             std::cout << "  Sum of buffer capacities: 0 (0 GB)\n";
